added some test
